refactor(ci): split setup-cypress into separate AWS and server scripts
Addresses review feedback on #5586: two entry points instead of one script with
a `--hosting-type` switch.

  scripts/setupCypress/shared.js               config write + Pulumi output read
  scripts/setupCypress/setupCypressAws.js      AWS entry
  scripts/setupCypress/setupCypressServer.js   self-hosted entry

The win beyond structure is that each script now takes only the arguments it
actually has: the AWS one keeps `--env`, `--project-folder` and `--localhost`
(all meaningless without a deployment), while the server one takes `--api-url`
and `--admin-url` and requires both. Previously every flag was accepted by one
script regardless of whether it applied.

`yarn setup-cypress` still runs the AWS script, so existing callers keep working
untouched - push.wac.ts, the AWS /e2e jobs, the frozen v5 Cypress workflow and
docs/CONTRIBUTING.md. `setup-cypress:aws` is added as an explicit alias, and
`setup-cypress:server` is what the new self-hosted job calls.

CONTRIBUTING.md gains a note pointing at the server variant, since the existing
instructions describe reading values from deployment state files - which a
self-hosted project does not have.

Verified: server happy path substitutes both URLs and blanks the Cognito values;
server without arguments exits with a usage example and leaves no
half-written config; the AWS project-folder guard still fires; both
`setup-cypress` and `setup-cypress:aws` resolve to the same script.
