+++ package.json
@@ -120,7 +120,9 @@
     "postinstall": "yarn node ./scripts/linkWorkspaces.js",
     "prepublishOnly": "tsx scripts/prepublishOnly/src/index.ts",
     "release": "tsx ./scripts/release/src/index.ts",
-    "setup-cypress": "tsx scripts/setupCypress",
+    "setup-cypress": "tsx scripts/setupCypress/setupCypressAws.js",
+    "setup-cypress:aws": "tsx scripts/setupCypress/setupCypressAws.js",
+    "setup-cypress:server": "tsx scripts/setupCypress/setupCypressServer.js",
     "setup-env-files": "node scripts/setupEnvFiles",
     "setup-project": "node scripts/setupProject",
     "system:build": "yarn && yarn ghawac build && yarn webiny sync-dependencies && yarn build",
